+++ test/mockfn/examples/basic_usage.clj
@@ -1,6 +1,7 @@
 (ns mockfn.examples.basic-usage
   (:require [clojure.test :refer :all]
             [mockfn.macros :as mfn]
+            [mockfn.mock :as mck]
             [mockfn.matchers :as matchers])
   (:import (clojure.lang ExceptionInfo)))
 
@@ -16,6 +17,12 @@
     (mfn/providing [(one-fn :argument-1) :result-1
                     (one-fn :argument-2) :result-2]
       (is (= :result-1 (one-fn :argument-1)))
+      (is (= :result-2 (one-fn :argument-2)))))
+
+  (testing "providing - cause function to throw exception"
+    (mfn/providing [(one-fn :argument-1) (mck/as-fn #(throw (ex-info "one-fn mocked to fail" {:arg %})))
+                    (one-fn :argument-2) :result-2]
+      (is (thrown? ExceptionInfo (one-fn :argument-1)))
       (is (= :result-2 (one-fn :argument-2)))))
 
   (testing "providing with more than one function"