+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project are documented in this file.
 
 ## Releases
 
+### 0.?.0 - 201?-00-00
+- Allow mocks to not only return base values, but invoke function calls.
+  This enables mocked functions to throw exceptions.
+
 ### 0.2.0 - 2018-08-20
 - Added syntax sugar for clojure.test.
 - Deprecated everything in the `mockfn.core` namespace.